requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
filelock==3.28.0
fsspec==2026.3.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai-whisper==20250625
regex==2026.4.4
requests==2.33.1
setuptools==81.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
yt-dlp==2026.3.17
